+++ pootle/apps/pootle_app/project_tree.py
@@ -10,58 +10,14 @@
 import errno
 import logging
 import os
-import re
 
 from django.conf import settings
 
 from pootle.core.log import STORE_RESURRECTED, store_log
 from pootle.core.utils.timezone import datetime_min
 from pootle_app.models.directory import Directory
-from pootle_language.models import Language
 from pootle_store.models import Store
-from pootle_store.util import absolute_real_path, relative_real_path
-
-
-#: Case insensitive match for language codes
-LANGCODE_RE = re.compile('^[a-z]{2,3}([_-][a-z]{2,3})?(@[a-z0-9]+)?$',
-                         re.IGNORECASE)
-#: Case insensitive match for language codes as postfix
-LANGCODE_POSTFIX_RE = re.compile(
-    '^.*?[-_.]([a-z]{2,3}([_-][a-z]{2,3})?(@[a-z0-9]+)?)$', re.IGNORECASE)
-
-
-def direct_language_match_filename(language_code, path_name):
-    name = os.path.splitext(os.path.basename(path_name))[0]
-    if name == language_code or name.lower() == language_code.lower():
-        return True
-
-    # Check file doesn't match another language.
-    if Language.objects.filter(code__iexact=name).count():
-        return False
-
-    detect = LANGCODE_POSTFIX_RE.split(name)
-    return (len(detect) > 1 and
-            (detect[1] == language_code or
-             detect[1].lower() == language_code.lower()))
-
-
-def match_template_filename(project, filename):
-    """Test if :param:`filename` might point at a template file for a given
-    :param:`project`.
-    """
-    ext = os.path.splitext(os.path.basename(filename))[1][1:]
-
-    # FIXME: is the test for matching extension redundant?
-    if ext in project.filetype_tool.template_extensions:
-        if ext not in project.filetype_tool.filetype_extensions:
-            # Template extension is distinct, surely file is a template.
-            return True
-        elif not find_lang_postfix(filename):
-            # File name can't possibly match any language, assume it is a
-            # template.
-            return True
-
-    return False
+from pootle_store.util import relative_real_path
 
 
 def get_matching_language_dirs(project_dir, language):
@@ -69,58 +25,43 @@ def get_matching_language_dirs(project_dir, language):
             if language.code == lang_dir]
 
 
-def get_non_existant_language_dir(project_dir, language, file_style,
-                                  make_dirs):
-    if file_style == "gnu":
-        return project_dir
-    elif make_dirs:
-        language_dir = os.path.join(project_dir, language.code)
-        os.mkdir(language_dir)
-        return language_dir
-    else:
-        raise IndexError("Directory not found for language %s, project %s" %
-                         (language.code, project_dir))
-
-
-def get_or_make_language_dir(project_dir, language, file_style, make_dirs):
+def get_or_make_language_dir(project_dir, language, make_dirs):
     matching_language_dirs = get_matching_language_dirs(project_dir, language)
-    if len(matching_language_dirs) == 0:
-        # If no matching directories can be found, check if it is a GNU-style
-        # project.
-        return get_non_existant_language_dir(project_dir, language, file_style,
-                                             make_dirs)
-    else:
+    if len(matching_language_dirs) != 0:
         return os.path.join(project_dir, matching_language_dirs[0])
 
+    if not make_dirs:
+        raise IndexError('Directory not found for language %s, project %s' %
+                         (language.code, project_dir))
 
-def get_language_dir(project_dir, language, file_style, make_dirs):
+    # If no matching directories can be found, create them
+    language_dir = os.path.join(project_dir, language.code)
+    os.mkdir(language_dir)
+    return language_dir
+
+
+def get_language_dir(project_dir, language, make_dirs):
     language_dir = os.path.join(project_dir, language.code)
     if not os.path.exists(language_dir):
-        return get_or_make_language_dir(project_dir, language, file_style,
-                                        make_dirs)
-    else:
-        return language_dir
+        return get_or_make_language_dir(project_dir, language, make_dirs)
+    return language_dir
 
 
-def get_translation_project_dir(language, project_dir, file_style,
-                                make_dirs=False):
+def get_translation_project_dir(language, project_dir, make_dirs=False):
     """Returns the base directory containing translations files for the
     project.
 
     :param make_dirs: if ``True``, project and language directories will be
                       created as necessary.
     """
-    if file_style == 'gnu':
-        return project_dir
-    else:
-        return get_language_dir(project_dir, language, file_style, make_dirs)
+    return get_language_dir(project_dir, language, make_dirs)
 
 
 def is_hidden_file(path):
     return path[0] == '.'
 
 
-def split_files_and_dirs(ignored_files, exts, real_dir, file_filter):
+def split_files_and_dirs(ignored_files, exts, real_dir):
     files = []
     dirs = []
     child_paths = [
@@ -134,7 +75,7 @@ def split_files_and_dirs(ignored_files, exts, real_dir, file_filter):
         should_include_file = (
             os.path.isfile(full_child_path)
             and os.path.splitext(full_child_path)[1][1:] in exts
-            and file_filter(full_child_path))
+        )
         if should_include_file:
             files.append(child_path)
         elif os.path.isdir(full_child_path):
@@ -213,11 +154,9 @@ def create_or_resurrect_dir(name, parent):
 
 
 # TODO: rename function or even rewrite it
-def add_files(translation_project, ignored_files, exts, relative_dir, db_dir,
-              file_filter=lambda _x: True):
+def add_files(translation_project, ignored_files, exts, relative_dir, db_dir):
     podir_path = to_podir_path(relative_dir)
-    files, dirs = split_files_and_dirs(
-        ignored_files, exts, podir_path, file_filter)
+    files, dirs = split_files_and_dirs(ignored_files, exts, podir_path)
     file_set = set(files)
     dir_set = set(dirs)
 
@@ -255,7 +194,7 @@ def add_files(translation_project, ignored_files, exts, relative_dir, db_dir,
             exts,
             fs_subdir,
             db_subdir,
-            file_filter)
+        )
         files += _files
         new_files += _new_files
         is_empty &= _is_empty
@@ -271,177 +210,19 @@ def to_podir_path(path):
     return os.path.join(settings.POOTLE_TRANSLATION_DIRECTORY, path)
 
 
-def find_lang_postfix(filename):
-    """Finds the language code at end of a filename."""
-    name = os.path.splitext(os.path.basename(filename))[0]
-    if LANGCODE_RE.match(name):
-        return name
-
-    match = LANGCODE_POSTFIX_RE.match(name)
-    if match:
-        return match.groups()[0]
-
-    for code in Language.objects.values_list('code', flat=True):
-        if (name.endswith('-'+code) or name.endswith('_'+code) or
-            name.endswith('.'+code) or
-            name.lower().endswith('-'+code.lower()) or
-            name.endswith('_'+code) or name.endswith('.'+code)):
-            return code
-
-
 def translation_project_dir_exists(language, project):
     """Tests if there are translation files corresponding to the given
     :param:`language` and :param:`project`.
     """
-    if project.get_treestyle() == "gnu":
-        # GNU style projects are tricky
-
-        if language.code == 'templates':
-            # Language is template look for template files
-            for dirpath_, dirnames, filenames in os.walk(
-                    project.get_real_path()):
-                for filename in filenames:
-                    if (project.file_belongs_to_project(filename,
-                                                        match_templates=True)
-                            and match_template_filename(project, filename)):
-                        return True
-        else:
-            # find files with the language name in the project dir
-            for dirpath_, dirnames, filenames in os.walk(
-                    project.get_real_path()):
-                for filename in filenames:
-                    # FIXME: don't reuse already used file
-                    if (project.file_belongs_to_project(filename,
-                                                        match_templates=False)
-                            and direct_language_match_filename(language.code,
-                                                               filename)):
-                        return True
-    else:
-        # find directory with the language name in the project dir
-        try:
-            dirpath_, dirnames, filename = os.walk(
-                project.get_real_path()).next()
-            if language.code in dirnames:
-                return True
-        except StopIteration:
-            pass
+    # find directory with the language name in the project dir
+    try:
+        dirpath_, dirnames, filename = os.walk(project.get_real_path()).next()
+        if language.code in dirnames:
+            return True
+    except StopIteration:
+        pass
 
     return False
-
-
-def get_translated_name_gnu(translation_project, store):
-    """Given a template :param:`store` and a :param:`translation_project` return
-    target filename.
-    """
-    pootle_path_parts = store.pootle_path.split('/')
-    pootle_path_parts[1] = translation_project.language.code
-    pootle_path = '/'.join(pootle_path_parts[:-1])
-    if not pootle_path.endswith('/'):
-        pootle_path = pootle_path + '/'
-
-    suffix = (
-        "%s%s%s"
-        % (translation_project.language.code,
-           os.extsep,
-           store.filetype.extension))
-    # try loading file first
-    try:
-        target_store = translation_project.stores.live().get(
-            parent__pootle_path=pootle_path,
-            name__iexact=suffix,
-        )
-        return (target_store.pootle_path,
-                target_store.file and target_store.file.path)
-    except Store.DoesNotExist:
-        target_store = None
-
-    # is this GNU-style with prefix?
-    use_prefix = (store.parent.child_stores.live().exclude(file="").count() > 1
-                  or translation_project.stores.live().exclude(
-                      name__iexact=suffix, file='').count())
-    if not use_prefix:
-        # let's make sure
-        for tp in translation_project.project.translationproject_set.exclude(
-                language__code='templates').iterator():
-            temp_suffix = (
-                "%s%s%s"
-                % (tp.language.code,
-                   os.extsep,
-                   store.filetype.template_extension))
-            if tp.stores.live().exclude(
-                    name__iexact=temp_suffix).exclude(file="").count():
-                use_prefix = True
-                break
-
-    if use_prefix:
-        if store.translation_project.language.code == 'templates':
-            tprefix = os.path.splitext(store.name)[0]
-            # FIXME: we should detect separator
-            prefix = tprefix + '-'
-        else:
-            prefix = os.path.splitext(store.name)[0][:-len(
-                store.translation_project.language.code)]
-            tprefix = prefix[:-1]
-
-        try:
-            target_store = translation_project.stores.live().filter(
-                parent__pootle_path=pootle_path,
-                name__in=[
-                    tprefix + '-' + suffix,
-                    tprefix + '_' + suffix,
-                    tprefix + '.' + suffix,
-                    tprefix + '-' + suffix.lower(),
-                    tprefix + '_' + suffix.lower(),
-                    tprefix + '.' + suffix.lower(),
-                ],
-            )[0]
-
-            return (target_store.pootle_path,
-                    target_store.file and target_store.file.path)
-        except (Store.DoesNotExist, IndexError):
-            pass
-    else:
-        prefix = ""
-
-    if store.file:
-        path_parts = store.file.path.split(os.sep)
-        name = prefix + suffix
-        path_parts[-1] = name
-        pootle_path_parts[-1] = name
-    else:
-        path_parts = store.parent.get_real_path().split(os.sep)
-        path_parts.append(store.name)
-
-    return '/'.join(pootle_path_parts), os.sep.join(path_parts)
-
-
-def get_translated_name(translation_project, store):
-    name = os.path.splitext(store.name)[0]
-
-    if store.file:
-        path_parts = store.file.name.split(os.sep)
-    else:
-        path_parts = store.parent.get_real_path().split(os.sep)
-        path_parts.append(store.name)
-
-    pootle_path_parts = store.pootle_path.split('/')
-
-    # Replace language code
-    path_parts[1] = translation_project.language.code
-    pootle_path_parts[1] = translation_project.language.code
-
-    # Replace extension
-    path_parts[-1] = (
-        "%s.%s"
-        % (name,
-           store.filetype.extension))
-    pootle_path_parts[-1] = (
-        "%s.%s"
-        % (name,
-           store.filetype.extension))
-
-    return ('/'.join(pootle_path_parts),
-            absolute_real_path(os.sep.join(path_parts)))
 
 
 def does_not_exist(path):