+++ pootle/static/js/shared/models/project.js
@@ -19,7 +19,6 @@ export const Project = Backbone.Model.extend({
     code: '',
     fullname: '',
     checkstyle: 'standard',
-    treestyle: 'auto',
     source_language: '',
     ignoredfiles: '',
     filetypes: [],