+++ pytest_pootle/fixtures/models/project.py
@@ -21,7 +21,6 @@ def _require_project(code, name, source_language, **kwargs):
         'fullname': name,
         'source_language': source_language,
         'checkstyle': 'standard',
-        'treestyle': 'auto',
     }
     criteria.update(kwargs)
     new_project = Project.objects.get_or_create(**criteria)[0]
@@ -87,7 +86,6 @@ def project0_directory(po_directory, project0):
 
 @pytest.fixture
 def project0_nongnu(project0_directory, project0, settings):
-    project0.treestyle = "nongnu"
     project0.save()
     project_dir = os.path.join(
         settings.POOTLE_TRANSLATION_DIRECTORY, project0.code)