+++ tests/models/project.py
@@ -1,12 +1,11 @@
 # -*- coding: utf-8 -*-
 #
 # Copyright (C) Pootle contributors.
+# Copyright (C) Zing contributors.
 #
-# This file is a part of the Pootle project. It is distributed under the GPL3
+# This file is a part of the Zing project. It is distributed under the GPL3
 # or later license. See the LICENSE file for a copy of the license and the
 # AUTHORS file for copyright and authorship information.
-
-import os
 
 import pytest
 
@@ -299,24 +298,6 @@ def test_root_hide_permissions(po_directory, nobody, default, admin, hide,
     assert items_equal(Project.accessible_by_user(nobody), ALL_PROJECTS)
     assert items_equal(Project.accessible_by_user(foo_user), ALL_PROJECTS)
     assert items_equal(Project.accessible_by_user(bar_user), ALL_PROJECTS)
-
-
-@pytest.mark.django_db
-def test_project_create_with_none_treestyle(english, templates, settings):
-    project = Project.objects.create(
-        code="foo",
-        fullname="bar",
-        source_language=english,
-        treestyle="none")
-    assert not os.path.exists(
-        os.path.join(
-            settings.POOTLE_TRANSLATION_DIRECTORY,
-            project.code))
-    project.save()
-    assert not os.path.exists(
-        os.path.join(
-            settings.POOTLE_TRANSLATION_DIRECTORY,
-            project.code))
 
 
 @pytest.mark.django_db